debian/control: Bump libgweather dependendency to rebuild against 40.0
libgweather had some internal API/ABI changes (not reflected by a soname
update as they apply at runtime only) so e-d-s needs to be rebuilt
against it.

As per this, bump the build dependency. Even though we supports both
versions and this is technically not required, we depend on a newer
release to ensure that we'll only rebuild this package when available.