Add patch to prevent infinite loop of memory allocation with GLib 2.73.x
When e-d-s converts flags to strings, it did so in a way that is not
robust against having a name for the absence of flags. GLib 2.73.x
in experimental added a name for the absence of TLS certificate flags,
causing e-d-s to go into an infinite loop allocating a byte at a time
until the OOM killer stops it.